+++ src/components/common/button/SearchHeaderButton.tsx
@@ -9,7 +9,7 @@ import {
   Text,
 } from 'react-native';
 import Icon from 'react-native-vector-icons/FontAwesome';
-import {RootStackParamList} from '../../router';
+import {RootStackParamList} from '../../../router';
 interface SearchHeaderButtonProps {}
 
 const SearchHeaderButton = ({}: SearchHeaderButtonProps) => {
