+++ src/components/signForm/SignTemplate.tsx
@@ -1,4 +1,4 @@
-import BackHeader from '@common/header/BackHeader';
+import BackHeader from '@components/common/header/BackHeader';
 import {colors} from '@styles/color';
 import React from 'react';
 import {View, Image, StyleSheet} from 'react-native';
