+++ src/pages/Sign/passwordPage/NewPasswordPage.tsx
@@ -2,12 +2,12 @@ import React, {useEffect} from 'react';
 import {colors} from '@styles/color';
 import {StyleSheet, Text, View} from 'react-native';
 import CommonButton from '../../../atoms/buttons/CommonButton';
-import SignInput from '@common/input/SignInput';
+import SignInput from '@components/common/input/SignInput';
 import {PasswordResetSchema, PasswordResetType} from '@utils/validation';
 import {useForm, Controller} from 'react-hook-form';
 import {zodResolver} from '@hookform/resolvers/zod';
 import {NewPasswordInput} from '@utils/sign';
-import BackHeader from '@common/header/BackHeader';
+import BackHeader from '@components/common/header/BackHeader';
 import {RouteProp, useNavigation, useRoute} from '@react-navigation/native';
 import {NativeStackNavigationProp} from '@react-navigation/native-stack';
 import {RootStackParamList} from '../../../router';
