+++ tsconfig.json
@@ -12,7 +12,7 @@
       "@pages/*": ["./src/pages/*"],
       "@apis/*": ["./src/apis/*"],
       "@interfaces/*": ["./src/interfaces/*"],
-      "@common/*": ["./src/common/*"]
+      "@common/*": ["src/components/common/*"]
     }
   }
 }
